refactoring of code by creating BusStopUtility so that codes can now be shared between MyStatus and RealtimeAccess instead of duplication
